refactor(BottomSheet): BottomSheet 컴포넌트 분리 및 SelectBox 폴더 위치 수정
- BottomSheet의 틀과 내부 폼 데이터를 각각 BottomSheetBase와 RestaurantForm으로 분리하여 폼을 주입하도록 함

- SelectBox는 common 폴더로 옮김
