+++ README.md
@@ -50,9 +50,9 @@ UI
 
 ### 리팩토링
 
+- [x] 컴포넌트 variant 설정
 - [ ] 컴포넌트 구조 재설계 및 적용
 - [ ] 상수화 적용
-- [x] 컴포넌트 variant 설정
 - [ ] CSS 분리
 
 ### 테스트 코드
